TP2/requirements.txt
# Dependencias principales para TP2
aiohttp
beautifulsoup4
lxml
Pillow
selenium
aiofiles
requests
webdriver-manager
